fix(db): repair fresh-install db:init on PostgreSQL (8 root causes)
A fresh `pnpm run db:init` (what a new ANTON Local install runs) failed on a
clean PostgreSQL: the consolidated schema + full migration replay collided in 8
places (17 incl. cascades). Existing installs were unaffected — their migration
ids are already recorded and skipped — which is why this stayed latent and kept
the E2E CI workflow red on every push.

Root causes fixed (all idempotent + safe for existing installs: already-applied
ids are skipped by the runner, and every edit is a guard / cast / CREATE IF NOT
EXISTS / no-op):
- 004, 047: columns folded into schema.postgresql.sql -> ADD COLUMN IF NOT EXISTS
- 077 community foundation: community_identity/connections/mail/group_members were
  only created by runtime DDL in community.ts (at app boot), never by a migration,
  so db:init's ALTERs failed -> CREATE them (IF NOT EXISTS) at the top of 077;
  unblocks 10 cascade migrations (079/080/085/086/089/100/102/208/220)
- 204 school: student_class_enrollments was only ever ALTER'd (old SQLite
  try/catch), never created -> CREATE IF NOT EXISTS before the ALTERs
- 056 markets: 3 GIN indexes referenced columns added later by 065 -> guard behind
  column-existence DO-blocks; round(double,int) -> round(x::numeric, n)
- 057 markets partitioning: the partition swap never succeeded anywhere (recorded
  applied, but tables stay non-partitioned on every real install; it drops columns
  063/070 and ~10 services need, and breaks single-column FKs) -> documented no-op
- 105: round(double precision, integer) -> ::numeric cast

Verified: real `pnpm run db:init` against a fresh throwaway DB completes with zero
[pg-migrations] FAILED (exit 0); db tests 90/90 pass.

+++ server/db/migrations-pg/105_fix_timestamp_columns.sql
@@ -25,8 +25,8 @@ SELECT
   prediction_type,
   COUNT(*) as total,
   SUM(CASE WHEN was_correct = 1 THEN 1 ELSE 0 END) as correct,
-  ROUND(AVG(CASE WHEN was_correct = 1 THEN 1.0 ELSE 0.0 END), 4) as accuracy,
-  ROUND(AVG(brier_score), 4) as avg_brier,
+  ROUND(AVG(CASE WHEN was_correct = 1 THEN 1.0 ELSE 0.0 END)::numeric, 4) as accuracy,
+  ROUND(AVG(brier_score)::numeric, 4) as avg_brier,
   MIN(validated_at) as first_validated,
   MAX(validated_at) as last_validated
 FROM market_predictions
